Update build process to work on ARM devices such as the Raspberry Pi
Add builder user to prevent issue where npm will not run scripts as root. Add ruby and install fpm, as the ARM version of fpm will be needed. Compress .deb package with gzip instead of xz, xz compression seems to make the .deb package corrupt. Clean up old output directory if it exists, and set 777 permissions on it so container user can write to it.